+++ src/index.js
@@ -1,11 +1,17 @@
 import Pool from './pool/Pool.js'
 import defaults from './lib/defaults.js'
 
-// ok guess we're gonna do this...
 const dumpster = function (opts) {
   opts = Object.assign({}, defaults, opts)
-  let pool = new Pool(opts)
-  pool.start()
+  const pool = new Pool(opts)
+  // start on the next tick, so listeners attached right after this call are in place
+  setImmediate(() => {
+    try {
+      pool.start()
+    } catch (err) {
+      pool.abort(err)
+    }
+  })
   return pool
 }
 export default dumpster