+++ src/worker/04-filter.js
@@ -1,12 +1,12 @@
-const filterPage = function (page) {
-  if (page.isRedirect) {
+const filterPage = function (page, opts) {
+  if (page.isRedirect && !opts.redirects) {
     return 'redirect'
-  } else if (page.isDisambig) {
+  } else if (page.isDisambig && !opts.disambiguation) {
     return 'disambig'
   } else if (page.isEmpty) {
     return 'empty'
   }
-  return true // keep all pages
+  return true // keep this page
 }
 
 export default filterPage